Revert "augmented the TravisCI post / coverage task using the code from the markdown-it-include plugin. Also fixed a (non-fatal) error in the Makefile."
This reverts the nyc/coverage edits done in the mainline repo.

TravisCI output for previous commit:

```
$ cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js
/home/travis/build/GerHobbelt/markdown-it-include/node_modules/coveralls/bin/coveralls.js:19
      throw err;
      ^
Bad response: 422 {"message":"Couldn't find a repository matching this job.","error":true}
(Use `node --trace-uncaught ...` to show where the exception was thrown)
```

so, as before, coveralls is not working for me. As `nyc` delivers the coverage info I need anyway, on *any* platform, I'll stick with that!